requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
google-auth
pyjwt
cryptography
stripe